requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# Yahoo Fantasy API
yfpy>=1.0.0
yahoo-oauth>=1.0.0

# AI API clients
openai>=1.0.0
anthropic>=0.7.0

# Data processing and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# HTTP and API utilities
aiohttp>=3.8.0
httpx>=0.24.0

# Data storage and serialization
pyyaml>=6.0
jsonschema>=4.17.0

# Logging and utilities
rich>=13.0.0
click>=8.1.0
tabulate>=0.9.0

# Development and testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
